requirements.txt
langchain==0.3.13
langchain-openai==0.3.1
langchain-core==0.3.63
langchain-community==0.3.13
langchain-text-splitters==0.3.8
boto3>=1.34.10
moto>=4.2.14
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=0.27.0
requests>=2.31.0